Update encoder interfaces and add interfaces for decoder
:Release Notes:
Update existing encoder interfaces and add new interfaces for
supporting decoder

:Detailed Notes:
Currently MCIL supports only encoding using HW. But we have to
extend this for decoder as well for all platforms.
So in this patch set we are modifying the existing types/functions
etc so that they can be used for decoding as well.
Also we have added new interface like VideoDecoderAPI to be able to
used for adding deocder implementation in codec-impl.

:Testing Performed:
Along with the change for V4l2 implementation done in codec_impl
for decoder and the existing modifications for encoder have tested
all the changes working fine in production and OSE platforms.

:QA Notes:
Test-cases: To be added

:Issues Addressed:
To be added

Change-Id: I8d8b00503d0c71952044d70b3183ef16b9371333
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/317934
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>