Fix interfaces for getting current input buffer id and split DoReset
:Release Notes:
Fix get current input buffer id and split do reset interface function

:Detailed Notes:
Currently GetCurrentInputBufferId returns flush buffer id if the
current input buffer pointer is null. But we should not process
the flush buffer if the id of current input buffer id is not -2.
So change the return type so that flase can be returned if current
input buffer pointer is null.
Also DoReset function solves two purposes and is called in two
phases. First to reset the current input buffer pointer and than
to reset the decoding buffers for actual flush to happen. But the
single function name creates confusion and difficult to understand.
So plit the function as ResetInputBuffer and ResetDecodingBuffers,
so that they can be implemented separately as per their purpose.

:Testing Performed:
All types of URL playbacks,MSE playback and WebCodec TC are working
same as before. Shaka player (Angel One, Big buck bunny) playing as in.
http://collab.lge.com/main/display/WEBOSDOCS/WebCodecs+Test+cases+and+Status
with corresponding changes in implementation part.

:QA Notes:
NA

:Issues Addressed:
NA

Change-Id: Ic4d96249e8be6b793c64b7fc811ac0eac5d4627a
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/326861
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>