Rename decoder/encoder delegate classes as client classes
:Release Notes:
Rename video decoder and encoder delegate classes as client class

:Detailed Notes:
Currently the parent class which initiate and control the decoder
implemenation is named as delegate class.
But the appropriate name would be as client class since we get the
input buffers and pass the output buffers to Client actually.

:Testing Performed:
Build is successfull there is no logical change other than name

:QA Notes:
NA

:Issues Addressed:
NA

Change-Id: I40953798d841012ce12c3a99b0bf39cadbbb772a
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/327253
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>