Remove OutputRecord type from mcil since not needed
:Release Notes:
Remove OutputRecord type from mcil since storing the record
data is not needed in mcil

:Detailed Notes:
Earlier we allowed to store the OutputRecord data in mcil to
ensure that the records are cleared after the destroy buffer is
successful. But we dont need the OutputRecord data in MCIL and
it is specific to Chromium.
Also since we can not store Chromium VideoFrame in output buffers
class for output so the destruction of DMA Buf FD are delayed till
next output frames are created.
By moving the OutputRecord to Chromium we can store the VideoFrame
associated with the DMA Buff FD and the same can be closed at the
right time when the output buffers are destroyed.

:Testing Performed:
All types of URL playbacks,MSE playback and WebCodec TC are working
same as before. Shaka player (Angel One, Big buck bunny) playing as in.
http://collab.lge.com/main/display/WEBOSDOCS/WebCodecs+Test+cases+and+Status

:QA Notes:
NA

:Issues Addressed:
NA

Change-Id: Iee8044289c98c190dc7810bda6c60d750b134359
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/328433
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>