Fix v4l2 encoder resolution change and freezing issue
:Release Notes:
In case of some resolution like 480x270 driver expects a diffrent
size of input buffer like 512x270. In that case input buffers has
to be scaled to 512x270.
We need to pass these new input_frame_size_ to client (Chromium).
So that client can pass the buffers with correct size for encoding.
Without this either it fails in client or in driver when we push
buffers with wrong size.
Hence we added input_frame_size to EncoderClientConfig to inform
the client about the same.
Also we added to fail and return false when we receive a resolution
not supported by the driver.
Below new H264 properties are added by following Chromium94 v4l2
encoder implementation for freezing issue in sites like MS Teams.
1. V4L2_CID_MPEG_VIDEO_H264_I_PERIOD
2. V4L2_CID_MPEG_VIDEO_H264_LOOP_FILTER_MODE
3. V4L2_CID_MPEG_VIDEO_H264_ENTROPY_MODE
4. V4L2_CID_MPEG_VIDEO_H264_8X8_TRANSFORM

:Testing Performed:
1. Start video call through MS Teams & join the call.
   Device A - Rpi4 with C270E
   Device B - Laptop
2. Local preview in Device A shoule be fine.
3. The Camera preview of Rpi4 should be fine in Device B(Laptop).

:QA Notes:
Test-cases: QWTC-7943

:Issues Addressed:
[WRO-15101] The green macro blocked data are streamed during video call

Change-Id: I26211033712c7fb980a24c7b99e4baf3661e2d38
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/338760
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: <kalaimani.k@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>