Fix MISRA C++-2008 Declarators static issues
:Release Notes:
Fix MISRA C++:2008 Rule 8-0-1, Rule 8-4-2, Rule 8-5-2, Rule 8-5-3.

:Detailed Notes:
Rule 8-4-2 is "identifiers used for the parameters in a redeclaration
of a function shall be identical to those in the declaration".
Solved 21 issues.
Rule 8-0-1 is "Init-declarator-list or a member-declarator-list shall
consist of a single init-declarator or member-declarator".
Solved 3 issues.
Rule 8-5-3 is "In an enumerator list, the = construct shall not be
used to explicitly initialize members other than the first, unless
all items are explicitly initialized" Solved 4 issues.
Rule 8-5-2 is "All variables shall have a defined value before they
are used". Solved 1 issue.

:Testing Performed:
QWTC-12386, QWTC-13015, QWTC-11386, QWTC-11385

:QA Notes:
N/A

:Issues Addressed:
[WRQ-29340] Fix 'MISRA C++ Declarators' static issues raised in MCIL

Change-Id: I44330e28c4935b766e0d2e851f8eb4c83edfbfff
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/427726
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Subhasish Mishra <subhasish.mishra@lge.com>
Reviewed-by: Praveen Kumar P <praveen.p@lge.com>
Reviewed-by: Subhasish Mishra <subhasish.mishra@lge.com>