Fix MISRA C++ 2008 Templates static issues
:Release Notes:
Fixed MISRA C++ 2008 Rule 14-6-1 static issues

:Detailed Notes:
Rule 14-6-1 is "In a class template with a dependent base,
any name that may be found in that dependent base shall be
referred to using a qualified-id or this->". Added scope resolution
operator or resolution to mcil namespace. Adjusted formatting.

:Testing Performed:
QWTC-12386, QWTC-13015, QWTC-11386, QWTC-11385, QWTC-11309

:QA Notes:
N/A

:Issues Addressed:
[WRQ-28770] Fix 'MISRA C++ Templates' static issues raised in MCIL

Change-Id: I870d351f477ba860ee12ef9a3a26855b2ac4abb6
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/429918
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Subhasish Mishra <subhasish.mishra@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Subhasish Mishra <subhasish.mishra@lge.com>