fixup! Fix MISRA C++ Lexical Conventions static issues
:Release Notes:
Fixed MISRA C++:2008 Rule 2-10-1 static issues.

:Detailed Notes:
Reverted the static issue fix for CID-10523424 as the change suggests
to modify the variable name of DecoderConfig::outpuMode because the
type of the variable is of the same name i.e. OutpuMode.
But as the media::WebOSVideoDecodeAccelerator of chromium uses this
data member as DecoderConfig::outpuMode, hence we need to keep the
variable name as it is in mcil::DecoderConfig.

:Testing Performed:
QWTC-12386, QWTC-13015, QWTC-11386, QWTC-11385, QWTC-11309

:QA Notes:
N/A

:Issues Addressed:
[WRQ-28775] Fix 'MISRA C++ Lexical Conventions' static issues MCIL

Change-Id: I1c9f2829cf39c90b02796c54988396d33b3b6837
Reviewed-on: http://gpro.lge.com/c/webosose/media-codec-interface/+/435603
Tested-by: Subhasish Mishra <subhasish.mishra@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Subhasish Mishra <subhasish.mishra@lge.com>